requirements.txt
langchain-core==0.3.21
python-dotenv==1.0.1
psycopg2-binary==2.9.10
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
uvicorn==0.32.1
langchain-cohere==0.3.3